+++ src/Compute/Compute/Extension/GetAzureVMExtensionCommand.cs
@@ -17,17 +17,23 @@ using Microsoft.Azure.Commands.Compute.Models;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using System.Management.Automation;
 using System.Linq;
+using System;
 
 namespace Microsoft.Azure.Commands.Compute
 {
-    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "VMExtension")]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "VMExtension", DefaultParameterSetName = GetExtensionParamSetName)]
     [OutputType(typeof(PSVirtualMachineExtension))]
     public class GetAzureVMExtensionCommand : VirtualMachineExtensionBaseCmdlet
     {
+
+        private const string GetExtensionParamSetName = "GetExtensionParameterSet",
+            VMParameterSetName = "VMParameterSet";
+
         [Parameter(
            Mandatory = true,
            Position = 0,
            ValueFromPipelineByPropertyName = true,
+           ParameterSetName = GetExtensionParamSetName,
            HelpMessage = "The resource group name.")]
         [ResourceGroupCompleter]
         [ValidateNotNullOrEmpty]
@@ -38,6 +44,7 @@ namespace Microsoft.Azure.Commands.Compute
             Mandatory = true,
             Position = 1,
             ValueFromPipelineByPropertyName = true,
+            ParameterSetName = GetExtensionParamSetName,
             HelpMessage = "The virtual machine name.")]
         [ResourceNameCompleter("Microsoft.Compute/virtualMachines", "ResourceGroupName")]
         [ValidateNotNullOrEmpty]
@@ -48,10 +55,18 @@ namespace Microsoft.Azure.Commands.Compute
             Mandatory = false,
             Position = 2,
             ValueFromPipelineByPropertyName = true,
+            ParameterSetName = GetExtensionParamSetName,
             HelpMessage = "The extension name.")]
         [ResourceNameCompleter("Microsoft.Compute/virtualMachines/extensions", "ResourceGroupName", "VMName")]
         [ValidateNotNullOrEmpty]
         public string Name { get; set; }
+
+        [Parameter(
+            ParameterSetName = VMParameterSetName,
+            ValueFromPipeline = true,
+            HelpMessage = "Specifies the virtual machine object the extension is on.")]
+        [ValidateNotNullOrEmpty]
+        public PSVirtualMachine VM { get; set; }
 
         [Parameter(
             Position = 3,
@@ -66,34 +81,57 @@ namespace Microsoft.Azure.Commands.Compute
 
             ExecuteClientAction(() =>
             {
+                string virtualMachineName = "";
+                string resourceGroup = "";
+
+                if (this.ParameterSetName.Equals(VMParameterSetName))
+                {
+                    virtualMachineName = this.VM.Name;
+                    if (this.VM.ResourceGroupName == null)
+                    {
+                        WriteError("The incoming virtual machine must have a 'resourceGroupName'.", this.VM);
+                    }
+                    resourceGroup = this.VM.ResourceGroupName;
+                }
+                else
+                {
+                    virtualMachineName = VMName;
+                    resourceGroup = ResourceGroupName;
+                }
+
                 if (!string.IsNullOrEmpty(Name))
                 {
                     if (Status.IsPresent)
                     {
-                        var result = this.VirtualMachineExtensionClient.GetWithInstanceView(this.ResourceGroupName, this.VMName, this.Name);
-                        WriteObject(result.ToPSVirtualMachineExtension(this.ResourceGroupName, this.VMName));
+                        var result = this.VirtualMachineExtensionClient.GetWithInstanceView(resourceGroup, virtualMachineName, this.Name);
+                        WriteObject(result.ToPSVirtualMachineExtension(resourceGroup, virtualMachineName));
                     }
                     else
                     {
-                        var result = this.VirtualMachineExtensionClient.GetWithHttpMessagesAsync(this.ResourceGroupName,
-                            this.VMName, this.Name).GetAwaiter().GetResult();
-                        WriteObject(result.ToPSVirtualMachineExtension(this.ResourceGroupName, this.VMName));
+                        var result = this.VirtualMachineExtensionClient.GetWithHttpMessagesAsync(resourceGroup,
+                            virtualMachineName, this.Name).GetAwaiter().GetResult();
+                        WriteObject(result.ToPSVirtualMachineExtension(resourceGroup, virtualMachineName));
                     }
                 }
                 else
                 {
                     if (Status.IsPresent)
                     {
-                        var result = this.VirtualMachineExtensionClient.ListWithInstanceView(this.ResourceGroupName, this.VMName).Body.Value;
-                        WriteObject(result.ToList().Select(t => t.ToPSVirtualMachineExtension(this.ResourceGroupName, this.VMName)), true);
+                        var result = this.VirtualMachineExtensionClient.ListWithInstanceView(resourceGroup, virtualMachineName).Body.Value;
+                        WriteObject(result.ToList().Select(t => t.ToPSVirtualMachineExtension(resourceGroup, virtualMachineName)), true);
                     }
                     else
                     {
-                        var result = this.VirtualMachineExtensionClient.ListWithHttpMessagesAsync(this.ResourceGroupName, this.VMName).GetAwaiter().GetResult().Body.Value;
-                        WriteObject(result.ToList().Select(t => t.ToPSVirtualMachineExtension(this.ResourceGroupName, this.VMName)), true);
+                        var result = this.VirtualMachineExtensionClient.ListWithHttpMessagesAsync(resourceGroup, virtualMachineName).GetAwaiter().GetResult().Body.Value;
+                        WriteObject(result.ToList().Select(t => t.ToPSVirtualMachineExtension(resourceGroup, virtualMachineName)), true);
                     }
                 }
             });
+        }
+
+        private void WriteError(string message, params object[] args)
+        {
+            base.WriteError(new ErrorRecord(new Exception(String.Format(message, args)), "Error", ErrorCategory.NotSpecified, null));
         }
     }
 }