+++ src/Compute/Compute/ChangeLog.md
@@ -30,6 +30,7 @@
     - `Remove-AzContainerService`
     - `Remove-AzContainerServiceAgentPoolProfile`
     - `Update-AzContainerService`
+* Added `VMParameterSet` parameter set to `Get-AzVMExtension` cmdlet. 
 
 ## Version 4.8.0
 * New parameter `VM` in new parameter set `VMParameterSet` added to `Get-AzVMDscExtensionStatus` and `Get-AzVMDscExtension` cmdlets.